changePacket to uint32_t

+++ client/include/GameClient.hh
@@ -61,18 +61,18 @@ namespace client {
         void gameLoop();
 
         void manageSpawnEntity(uint32_t tick, uint32_t eventId, const std::string &spriteName,
-                               uint16_t entityId, int16_t pos_x, int16_t pos_y, int16_t hp);
+                               uint32_t entityId, int16_t pos_x, int16_t pos_y, int16_t hp);
 
-        void manageUpdateEntity(uint32_t tick, uint32_t eventId, uint16_t entityId, int16_t hp);
+        void manageUpdateEntity(uint32_t tick, uint32_t eventId, uint32_t entityId, int16_t hp);
 
-        void manageMoveEntity(uint32_t tick, uint32_t eventId, uint16_t entityId,
+        void manageMoveEntity(uint32_t tick, uint32_t eventId, uint32_t entityId,
                               int16_t vecx, int16_t vecy, int16_t posx, int16_t posy);
 
-        void manageDeleteEntity(uint32_t tick, uint32_t eventId, uint16_t entityId);
+        void manageDeleteEntity(uint32_t tick, uint32_t eventId, uint32_t entityId);
 
         void manageSyncro(uint32_t tick, int64_t time);
 
-        void managePlayerData(uint16_t playerId, uint8_t nbAttackPlayer);
+        void managePlayerData(uint32_t playerId, uint8_t nbAttackPlayer);
 
         void manageGameData();
